分页实现调整

+++ springboot-starter/src/main/java/com/wei/springboot/starter/bean/Page.java
@@ -16,9 +16,9 @@ public class Page<T> {
 
     private List<T> list = Collections.emptyList();
 
-    private long page = 1;
+    private int page = 1;
 
-    private long size = 20;
+    private int size = 20;
 
     private long total = 0;
